requirements.txt
python-dotenv
requests
gradio
retrying